feat: move metrics out of ethers-prometheus
