+++ rust/Dockerfile
@@ -17,6 +17,7 @@ COPY rust/main/agents rust/main/agents
 COPY rust/main/chains rust/main/chains
 COPY rust/main/hyperlane-base rust/main/hyperlane-base
 COPY rust/main/hyperlane-core rust/main/hyperlane-core
+COPY rust/main/hyperlane-metric rust/main/hyperlane-metric
 COPY rust/main/hyperlane-test rust/main/hyperlane-test
 COPY rust/main/ethers-prometheus rust/main/ethers-prometheus
 COPY rust/main/utils rust/main/utils
