+++ rust/main/chains/hyperlane-ethereum/Cargo.toml
@@ -31,6 +31,7 @@ tracing.workspace = true
 url.workspace = true
 
 hyperlane-core = { path = "../../hyperlane-core", features = ["async"] }
+hyperlane-metric = { path = "../../hyperlane-metric" }
 ethers-prometheus = { path = "../../ethers-prometheus", features = ["serde"] }
 
 [build-dependencies]
