+++ rust/main/chains/hyperlane-ethereum/src/rpc_clients/retrying.rs
@@ -3,9 +3,8 @@ use std::{fmt::Debug, str::FromStr, time::Duration};
 use crate::rpc_clients::{categorize_client_response, CategorizedResponse};
 use async_trait::async_trait;
 use ethers::providers::{Http, JsonRpcClient, ProviderError};
-use ethers_prometheus::json_rpc_client::{
-    PrometheusJsonRpcClient, PrometheusJsonRpcClientConfigExt,
-};
+use ethers_prometheus::json_rpc_client::PrometheusJsonRpcClient;
+use hyperlane_metric::prometheus_metric::PrometheusJsonRpcClientConfigExt;
 use serde::{de::DeserializeOwned, Serialize};
 use serde_json::Value;
 use thiserror::Error;
