+++ rust/main/hyperlane-base/Cargo.toml
@@ -47,6 +47,7 @@ backtrace = { workspace = true, optional = true }
 backtrace-oneline = { path = "../utils/backtrace-oneline", optional = true }
 
 ethers-prometheus = { path = "../ethers-prometheus", features = ["serde"] }
+hyperlane-metric = { path = "../hyperlane-metric" }
 hyperlane-core = { path = "../hyperlane-core", features = ["agent", "float"] }
 hyperlane-ethereum = { path = "../chains/hyperlane-ethereum" }
 hyperlane-fuel = { path = "../chains/hyperlane-fuel" }
