+++ rust/main/hyperlane-base/src/metrics/core.rs
@@ -13,7 +13,8 @@ use prometheus::{
 };
 use tokio::sync::RwLock;
 
-use ethers_prometheus::{json_rpc_client::JsonRpcClientMetrics, middleware::MiddlewareMetrics};
+use ethers_prometheus::middleware::MiddlewareMetrics;
+use hyperlane_metric::prometheus_metric::JsonRpcClientMetrics;
 
 use crate::metrics::{
     json_rpc_client::create_json_rpc_client_metrics, provider::create_provider_metrics,
