+++ rust/main/hyperlane-metric/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ethers-prometheus"
+name = "hyperlane-metric"
 documentation.workspace = true
 edition.workspace = true
 homepage.workspace = true
@@ -18,19 +18,10 @@ log.workspace = true
 maplit.workspace = true
 parking_lot.workspace = true
 prometheus.workspace = true
-serde = { workspace = true, features = ["derive"], optional = true }
+serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 static_assertions.workspace = true
 tokio = { workspace = true, features = ["time", "sync", "parking_lot"] }
+url.workspace = true
 
-# enable feature for this crate that is imported by ethers-rs
-primitive-types = { workspace = true, features = ["fp-conversion"] }
 hyperlane-core = { path = "../hyperlane-core", features = ["agent", "float"] }
-hyperlane-metric = { path = "../hyperlane-metric" }
-
-[build-dependencies]
-abigen = { path = "../utils/abigen", features = ["ethers"] }
-
-[features]
-default = []
-serde = ["dep:serde"]
