+++ rust/main/hyperlane-metric/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod prometheus_metric;
+pub mod utils;
